Updated all int to NSInteger for better 64 bit compatibility

+++ mediation/mediation/config/model/PubnativePlacementModel.h
@@ -17,6 +17,6 @@
 @property(nonatomic,strong)NSArray<PubnativePriorityRuleModel*>     *priority_rules;
 @property(nonatomic,strong)PubnativeDeliveryRuleModel               *delivery_rule;
 
-- (PubnativePriorityRuleModel*)priorityRuleWithIndex:(int)index;
+- (PubnativePriorityRuleModel*)priorityRuleWithIndex:(NSInteger)index;
 
 @end
